+++ mypy/main.py
@@ -545,6 +545,9 @@ def process_options(args: List[str],
         '--default-return', default=False, dest="default_return",
         help="Assume implicit default return type of None",
         group=based_group)
+    add_invertible_flag(
+        '--ignore-any-from-error', default=False, dest="ignore_any_from_error",
+        help="Don't include Any type from errors in no-any-expr messages.", group=based_group)
 
     config_group = parser.add_argument_group(
         title='Config file',