+++ mypy/options.py
@@ -122,6 +122,7 @@ class Options:
         self.auto_baseline = True
         self.default_return = False
         self.targets: List[str] = []
+        self.ignore_any_from_error = True
 
         # disallow_any options
         self.disallow_any_generics = flip_if_not_based(True)