+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- `ignore_any_from_errors` option to suppress `no-any-expr` messages from other errors
 
 ## [1.3.0]
 ### Added